Fix qwen_image_edit_2509_photous: use bfloat16 as default dtype
LoRA weights are stored in bfloat16; loading them into a float32 model
then fusing leaves some weights in bfloat16, causing a dtype mismatch
at runtime.

+++ qwen_image_edit_2509_photous/pytorch/loader.py
@@ -72,7 +72,7 @@ class ModelLoader(ForgeModel):
         Returns:
             QwenImageTransformer2DModel instance with LoRA weights merged.
         """
-        dtype = dtype_override if dtype_override is not None else torch.float32
+        dtype = dtype_override if dtype_override is not None else torch.bfloat16
 
         pipeline = QwenImageEditPipeline.from_pretrained(
             BASE_MODEL_ID,
@@ -93,7 +93,7 @@ class ModelLoader(ForgeModel):
 
         Returns a dict matching QwenImageTransformer2DModel.forward() signature.
         """
-        dtype = kwargs.get("dtype_override", torch.float32)
+        dtype = kwargs.get("dtype_override", torch.bfloat16)
         batch_size = kwargs.get("batch_size", 1)
 
         # From model config: in_channels=64 (img_in linear input dimension)
